Release Candidate 2
Fixes up an error in spawning on Level EX, changes the title of the credits placeholder, and adds the credits, along with a small tweak here and there.
At this point, the game is done. C'est fini, it's over, etc. This is Release Candidate 2, but I'm almost confident that this could be our gold release. Almost. There could be a lingering bug or 5 still hanging about. Who knows, not me right now.